+++ src/types.rs
@@ -16,6 +16,7 @@ use crate::{
     encryption::EncryptionAlgorithm,
     identifiers::{DeviceKeyId, UserId},
     impl_from_to_inner,
+    responses::OlmEncryptionInfo,
     vodozemac::Ed25519Signature,
 };
 
@@ -359,6 +360,172 @@ impl From<&RoomSettings> for matrix_sdk_crypto::store::RoomSettings {
             session_rotation_period_messages: value
                 .session_rotation_period_messages
                 .map(|count| count as usize),
+        }
+    }
+}
+
+/// The type of a {@link ProcessedToDeviceEvent}.
+#[wasm_bindgen]
+#[derive(Debug, Clone)]
+pub enum ProcessedToDeviceEventType {
+    /// A successfully-decrypted encrypted to-device message.
+    Decrypted,
+
+    /// An encrypted to-device message which could not be decrypted.
+    UnableToDecrypt,
+
+    /// An unencrypted to-device message (sent in clear).
+    PlainText,
+
+    /// An invalid to-device message that was ignored because it is missing some
+    /// required information to be processed (like no event `type` for
+    /// example)
+    Invalid,
+}
+
+/// Represent a to-device event after it has been decrypted by {@link
+/// OlmMachine#receiveSyncChanges}.
+#[wasm_bindgen(getter_with_clone)]
+#[derive(Debug, Clone)]
+pub struct DecryptedToDeviceEvent {
+    /// A JSON-encoded object containing the processed event.
+    /// The decrypted message as if it was sent in clear (For room keys or
+    /// secrets some part of the content might have been zeroize'd).
+    /// Represents a Matrix to-device message with `type`, `sender` and
+    /// `content` fields.
+    #[wasm_bindgen(readonly, js_name = "decryptedRawEvent")]
+    pub decrypted_raw_event: JsString,
+
+    #[wasm_bindgen(readonly, js_name = "encryptionInfo")]
+    /// The olm encryption information for the event.
+    pub encryption_info: OlmEncryptionInfo,
+}
+
+#[wasm_bindgen]
+impl DecryptedToDeviceEvent {
+    /// The type of processed to-device event. Always {@link
+    /// ProcessedToDeviceEventType.Decrypted} for this type.
+    #[wasm_bindgen(getter, js_name = "type")]
+    pub fn processed_type(&self) -> ProcessedToDeviceEventType {
+        ProcessedToDeviceEventType::Decrypted
+    }
+
+    /// Create a new `DecryptedToDeviceEvent`.
+    #[wasm_bindgen(constructor)]
+    pub fn new(decrypted_raw_event: JsString, encryption_info: OlmEncryptionInfo) -> Self {
+        Self { decrypted_raw_event, encryption_info }
+    }
+}
+
+/// Represent a ToDevice event sent in clear after it has been processed {@link
+/// OlmMachine#receiveSyncChanges}.
+#[wasm_bindgen(getter_with_clone)]
+#[derive(Debug, Clone)]
+pub struct PlainTextToDeviceEvent {
+    /// A JSON-encoded object containing the Matrix to-device message with
+    /// `type`, `sender` and `content` fields.
+    #[wasm_bindgen(readonly, js_name = "rawEvent")]
+    pub raw_event: JsString,
+}
+
+#[wasm_bindgen]
+impl PlainTextToDeviceEvent {
+    /// The type of processed to-device event. Always {@link
+    /// ProcessedToDeviceEventType.PlainText} for this type.
+    #[wasm_bindgen(getter, js_name = "type")]
+    pub fn processed_type(&self) -> ProcessedToDeviceEventType {
+        ProcessedToDeviceEventType::PlainText
+    }
+
+    /// Create a new `PlainTextToDeviceEvent`.
+    #[wasm_bindgen(constructor)]
+    pub fn new(raw_event: JsString) -> Self {
+        Self { raw_event }
+    }
+}
+
+/// Represent a ToDevice event that could not be decrypted after it has been
+/// processed.
+#[wasm_bindgen(getter_with_clone)]
+#[derive(Debug, Clone)]
+pub struct UTDToDeviceEvent {
+    /// A JSON-encoded object containing the original message of type
+    /// `m.room.encrypted` that failed to be decrypted.
+    #[wasm_bindgen(readonly, js_name = "wireEvent")]
+    pub wire_event: JsString,
+    // TODO: Add some OlmError in the future
+}
+
+#[wasm_bindgen]
+impl UTDToDeviceEvent {
+    /// The type of processed to-device event. Always {@link
+    /// ProcessedToDeviceEventType.UnableToDecrypt} for this type.
+    #[wasm_bindgen(getter, js_name = "type")]
+    pub fn processed_type(&self) -> ProcessedToDeviceEventType {
+        ProcessedToDeviceEventType::UnableToDecrypt
+    }
+
+    /// Create a new `UTDToDeviceEvent`.
+    #[wasm_bindgen(constructor)]
+    pub fn new(wire_event: JsString) -> Self {
+        Self { wire_event }
+    }
+}
+
+/// Represent an invalid ToDevice event that was ignored because it is missing
+/// some mandatory fields for example.
+#[wasm_bindgen(getter_with_clone)]
+#[derive(Debug, Clone)]
+pub struct InvalidToDeviceEvent {
+    /// A JSON-encoded object containing the original message as received from
+    /// sync.
+    #[wasm_bindgen(readonly, js_name = "wireEvent")]
+    pub wire_event: JsString,
+    // TODO: Add some error information here?
+}
+
+#[wasm_bindgen]
+impl InvalidToDeviceEvent {
+    /// The type of processed to-device event. Always {@link
+    /// ProcessedToDeviceEventType.Invalid} for this type.
+    #[wasm_bindgen(getter, js_name = "type")]
+    pub fn processed_type(&self) -> ProcessedToDeviceEventType {
+        ProcessedToDeviceEventType::Invalid
+    }
+
+    /// Create a new `InvalidToDeviceEvent`.
+    #[wasm_bindgen(constructor)]
+    pub fn new(wire_event: JsString) -> Self {
+        Self { wire_event }
+    }
+}
+
+/// Convert an `ProcessedToDeviceEvent` into a `JsValue`, ready to return to
+/// JavaScript.
+///
+/// JavaScript has no complex enums like Rust. To return structs of
+/// different types, we have no choice that hiding everything behind a
+/// `JsValue`.
+pub fn processed_to_device_event_to_js_value(
+    processed_to_device_event: matrix_sdk_crypto::types::ProcessedToDeviceEvent,
+) -> JsValue {
+    match processed_to_device_event {
+        matrix_sdk_crypto::types::ProcessedToDeviceEvent::Decrypted { raw, encryption_info } => {
+            JsValue::from(DecryptedToDeviceEvent::new(
+                raw.json().get().into(),
+                encryption_info
+                    .try_into()
+                    .expect("to-device encryption info will always convert to OlmEncryptionInfo"),
+            ))
+        }
+        matrix_sdk_crypto::types::ProcessedToDeviceEvent::UnableToDecrypt(utd) => {
+            JsValue::from(UTDToDeviceEvent::new(utd.json().get().into()))
+        }
+        matrix_sdk_crypto::types::ProcessedToDeviceEvent::PlainText(plain) => {
+            JsValue::from(PlainTextToDeviceEvent::new(plain.json().get().into()))
+        }
+        matrix_sdk_crypto::types::ProcessedToDeviceEvent::Invalid(invalid) => {
+            JsValue::from(InvalidToDeviceEvent::new(invalid.json().get().into()))
         }
     }
 }