+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # UNRELEASED
 
+-   Update matrix-rusk-sdk to `0f73ffde`, which includes:
+
+    -   feat(crypto): Add EncryptionInfo to Decrypted to-device variant ([#5074](https://github.com/matrix-org/matrix-rust-sdk/pull/5074))
+    -   Add variants for plain text and encrypted to-device events ([#4935](https://github.com/matrix-org/matrix-rust-sdk/pull/4935))
+
+-   **BREAKING**: `OlmMachine.receiveSyncChanges` now returns a list of
+    `ProcessedToDeviceEvent` instead of a JSON-encoded list of JSON-encoded events.
+    This allows making the difference between an event that was sent in clear and
+    the same event successfully decrypted.
+
 # matrix-sdk-crypto-wasm v14.2.0
 
 -   Log warnings when we fail to parse a backed-up room key
@@ -7,7 +17,7 @@
 
 # matrix-sdk-crypto-wasm v14.1.0
 
--   Update matrix-rusk-sdk to `0.11.0`, which includees:
+-   Update matrix-rusk-sdk to `0.11.0`, which includes:
 
     -   Add support for the shared history flag defined in
         [MSC3061](https://github.com/matrix-org/matrix-spec-proposals/pull/3061).